refactor(tabs): move the ViewTab type out of the app root
lib/openTabs persists a ViewTab and imported the type from App.tsx to describe
its own payload.

With this and the kinds move, nothing in lib depends on anything above it,
which is what the service layer needs before it can be extracted into a package.
